requirements/base.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-cleanup==4.0.0

django-imagekit==4.0.2
django-sendgrid-v5==0.8.0
django-widgets-improved==1.5.0
future==0.18.0
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
pilkit==2.0
Pillow==6.2.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0

six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
